requirements.txt
requests==2.31.0
python-dotenv==1.0.0
discord-webhook==1.2.1
jmespath>=1.0.1
